increment parents subdomainCount on registering domain
