+++ x/registry/keeper/register_domain.go
@@ -1,105 +1,12 @@
 package keeper
 
 import (
-	"errors"
-	"fmt"
 	incentivestypes "github.com/mycel-domain/mycel/x/incentives/types"
 	"github.com/mycel-domain/mycel/x/registry/types"
 	"strconv"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
-
-// Get is domain already taken
-func (k Keeper) GetIsDomainAlreadyTaken(ctx sdk.Context, domain types.Domain) (isDomainAlreadyTaken bool) {
-	_, isDomainAlreadyTaken = k.GetDomain(ctx, domain.Name, domain.Parent)
-	return isDomainAlreadyTaken
-}
-
-// Get is parent domain exist
-func (k Keeper) GetIsParentDomainExist(ctx sdk.Context, domain types.Domain) (isParentDomainExist bool) {
-	name, parent := domain.ParseParent()
-	_, isParentDomainExist = k.GetDomain(ctx, name, parent)
-	return isParentDomainExist
-}
-
-// Validate TLD registration
-func (k Keeper) ValidateRegisterTLD(ctx sdk.Context, domain types.Domain) (err error) {
-	if domain.Parent != "" {
-		err = sdkerrors.Wrapf(errors.New(domain.Parent),
-			types.ErrParentDomainMustBeEmpty.Error())
-	}
-	// TODO: Is Staked enough to register TLD
-	return err
-}
-
-// Validate SLD registration
-func (k Keeper) ValidateRegsiterSLD(ctx sdk.Context, domain types.Domain) (err error) {
-	isParentDomainExist := k.GetIsParentDomainExist(ctx, domain)
-	if !isParentDomainExist {
-		err = sdkerrors.Wrapf(errors.New(domain.Parent),
-			types.ErrParentDomainDoesNotExist.Error())
-	}
-
-	return err
-}
-
-// Validate subdomain GetRegistrationFee
-func (k Keeper) ValidateRegsiterSubdomain(ctx sdk.Context, domain types.Domain) (err error) {
-	isParentDomainExist := k.GetIsParentDomainExist(ctx, domain)
-	if !isParentDomainExist {
-		err = sdkerrors.Wrapf(errors.New(domain.Parent),
-			types.ErrParentDomainDoesNotExist.Error())
-	}
-	return err
-}
-
-// Validate domain
-func (k Keeper) ValidateDomain(ctx sdk.Context, domain types.Domain) (err error) {
-	// Type check
-	err = domain.Validate()
-	if err != nil {
-		return err
-	}
-	// Check if domain is already taken
-	isDomainAlreadyTaken := k.GetIsDomainAlreadyTaken(ctx, domain)
-	if isDomainAlreadyTaken {
-		err = sdkerrors.Wrapf(errors.New(fmt.Sprintf("%s.%s", domain.Name, domain.Parent)),
-			types.ErrDomainIsAlreadyTaken.Error())
-		return err
-	}
-
-	domainLevel := domain.GetDomainLevel()
-	switch domainLevel {
-	case 1: // TLD
-		// Validate TLD
-		err = k.ValidateRegisterTLD(ctx, domain)
-		if err != nil {
-			return err
-		}
-	case 2: // TLD
-		// Validate SLD
-		err = k.ValidateRegsiterSLD(ctx, domain)
-		if err != nil {
-			return err
-		}
-	default: // Subdomain
-		// Validate Subdomain
-		err = k.ValidateRegsiterSubdomain(ctx, domain)
-		if err != nil {
-			return err
-		}
-	}
-
-	return err
-}
-
-// Pay SLD registration fee
-func (k Keeper) PayTLDRegstrationFee(ctx sdk.Context, payer sdk.AccAddress, domain types.Domain, registrationPeriodInWeek uint) (err error) {
-	// TODO: Pay TLD registration fee
-	return err
-}
 
 // Pay SLD registration fee
 func (k Keeper) PaySLDRegstrationFee(ctx sdk.Context, payer sdk.AccAddress, domain types.Domain, registrationPeriodInWeek uint) (err error) {
@@ -119,6 +26,23 @@ func (k Keeper) AppendToOwnedDomain(ctx sdk.Context, owner string, name string, 
 	}
 }
 
+func (k Keeper) IncrementParentsSubdomainCount(ctx sdk.Context, domain types.Domain) {
+	// Check if domain is not TLD
+	level := domain.GetDomainLevel()
+	if level == 1 {
+		panic("domain is TLD")
+	}
+
+	// Increment parent's subdomain count
+	parentsName, parentsParent := domain.ParseParent()
+	parentDomain, found := k.GetDomain(ctx, parentsName, parentsParent)
+	if !found {
+		panic("parent not found")
+	}
+	parentDomain.SubdomainCount++
+	k.SetDomain(ctx, parentDomain)
+}
+
 func (k Keeper) RegisterDomain(ctx sdk.Context, domain types.Domain, owner sdk.AccAddress, registrationPeriodInWeek uint) (err error) {
 	// Validate domain
 	err = k.ValidateDomain(ctx, domain)
@@ -130,18 +54,19 @@ func (k Keeper) RegisterDomain(ctx sdk.Context, domain types.Domain, owner sdk.A
 	domainLevel := domain.GetDomainLevel()
 	switch domainLevel {
 	case 1: // TLD
-		// Pay TLD registration fee
-		err = k.PayTLDRegstrationFee(ctx, owner, domain, registrationPeriodInWeek)
-		if err != nil {
-			return err
-		}
-	case 2: // TLD
+		// TODO: Register TLD
+		return
+	case 2: // SLD
+		// Increment parents subdomain SubdomainCount
+		k.IncrementParentsSubdomainCount(ctx, domain)
 		// Pay SLD registration fee
 		err = k.PaySLDRegstrationFee(ctx, owner, domain, registrationPeriodInWeek)
 		if err != nil {
 			return err
 		}
 	default: // Subdomain
+		// Increment parents subdomain SubdomainCount
+		k.IncrementParentsSubdomainCount(ctx, domain)
 	}
 
 	// Append to owned domain
